Move MWStreams and Track6D classes into their own files
